+++ lib/util/remove.test.ts
